feat: restructure prisma imports and add prisma schema; update tsconfig include paths
