--- apps/api/src/prisma/prisma.ts
@@ -1,5 +1,0 @@
-import { PrismaClient } from 'prisma-client-7a5d7cf047f46244bd13fa0ec2311c64db7532942d10ac286efb5d925a2a8be0';
-
-const prisma = new PrismaClient();
-
-export default prisma;
