+++ apps/api/tsconfig.json
@@ -13,6 +13,6 @@
     }
   },
   "include": ["src/**/*",
-    "src/prisma/**/*"
+    "prisma/**/*"
   ]
 }
